requirements.txt
aiohttp==3.8.5
alembic==1.12.0
psycopg2==2.9.7
sqlalchemy==2.0.21
asyncpg==0.28.0
python-dotenv==1.0.0
fastapi[all]==0.95.0
pyTelegramBotAPI==4.14.0